Remove ngTouch
DEPRECATION NOTICE: Beginning with Angular 1.5, this directive is
deprecated and by default disabled. The directive will receive no
further support and might be removed from future releases. If you need
the directive, you can enable it with the
$touchProvider#ngClickOverrideEnabled function. We also recommend that
you migrate to FastClick. To learn more about the 300ms delay, this
Telerik article gives a good overview.