Fix #4355 - "1 hour" -> "24 hours"
Change the verification message from 1 to 24 hours

When you add a custom domain, after setting the verification TXT record
on your registrar, and clicking Verify in Settings>Domains you get a
message that says it may take up to 1 hour for the settings to be
updated. This should be changed to up to 24 hours, because users are
contacting us, and by the time we check their message, the domain has
already been verified.