Hotfix #4860 - See description
i would like to keep the Add User button, but instead give this message
in an error popup modal. this way users always see the Add User option.
This is also not quite true message in this case. I also need to enable
multi-member support first for this account. In fact, I will say that
Add Member is always present and we change the message based on whether
custom domains are there, and whether or not multi-member support is
enabled.