Fix #6089 ~ Hide native clear and fix infinite loop IE11 :bug:
If the listener was on the input (e:input) idk why but there was an infinite loop for
this input listener. As if the browser was emitting the input event itself. Weird.
Via the container itself, no more issue.
The native clear does not trigger a reset event nor an input it seems. Better to hide it
as we have one which works and looks better :)