Fix #7268  Falling back on the key api if the address is disabled (#7336)
* First step

https://github.com/ProtonMail/Angular/blob/v3/src/app/message/factories/keyCache.js#L74 can try to fallback to the keys api.

Then sendPreferences need to be adjusted. We want it to return primaryPinned = false in this case.
https://github.com/ProtonMail/Angular/blob/v3/src/app/composer/factories/sendPreferences.js#L94

I would change isOwnAddress to exclude the fallback addresses (e.g. look if the fetched keys from the keyCache actually has keys while the addressModel says the address doesn't allow receiving)
https://github.com/ProtonMail/Angular/blob/v3/src/app/composer/factories/sendPreferences.js#L32

Then we can add a function isFallbackAddress which does something similar but check if it's a fallback address. In that case can return primaryPinned = false.

* Add new method: getByEmail

* Review

* Review

* Use normalizeEmail string helper

* Address last part with @kaylukas comments

* Review

* Create default contacts

* Review

* Review

* Useless line

* Hotfix isOwnAddress call

* Move address helper and add isOwnAddress

* Use it

* Review isOwnAddress in contactItem directive

* Make getUserAddressesKeys public

* Add unit test for address helper

* Use isOwnAddress from address helper

* Add comments

* Address @kaylukas comments