Fix display of arrows for scrolling left/right (#7633)
When there is no scroll on `.settingsDashboard-plans`, `element.scrollWidth` (its "full" width) and `element.clientWidth` (its "displayed" width = the visible part on client between scrolls) are equals.

So `container.get(0).scrollWidth < container.get(0).clientWidth` was always set to false, and so arrows for scrolling were always displayed (they are hidden when button are disabled). 

A simple `<=` comparison instead of a `<` fixes it.