Fix #7843 - Allow ECC key generation (#8352)
* Add default encryption config

* Use default encryption config when we generate keys

* Member modal

* Setup organization

* generate organization

* generate key process

* Review

* Review

* Review key generation

Remove config choice in org and member modal.

* Review

* Warn the user that such a key already exists and discourages the user from generating a new key

* Review export on keyAlgorithm

* Review import

* Hotfix - Encryption config choice

* Apply logic described by @sanjanarajan to display algo name

EDDSA is the signature algorithm, ECDH is the encryption algorithm, ed25519 is the elliptic curve used for signing, curve25519 is the elliptic curve used for encryption. So this key can go by many names :)

If you use pmcrypto's keyInfo on the key, you will get a curve value of ed25519 and algorithm value of eddsa. Maybe you could just map ECDSA, EDDSA, and ECDH to "ECC" and then this can be described as "ECC (ed25519)".

* Hotfix - When reseting password, the newly created default key was RSA 2048-bit instead of X25519

Report from @emnproton

* Update description for encryption config