Add StyledPayPalButton component
To enable the use of a PayPal button that mimics the branding
styles outlined by PayPal, the StyledPayPalButton component
has been created. This component simply wraps PayPalButton
conveniently providing styles.

The styles are applied using the .paypal-button class which
configures the button css variables. This is all that needs
to be done since the button should look the same, regardless
of themes.

CP-1975