Add X-PM-PROTON-REPLY field when replying to Proton invites. Process it on organizer mode
Proton-to-Proton invites should never be updated by clients. While in principle
clients should be able to identify when the event has been updated already API-side,
as an extra measure it's good to have some identification of a Proton-to-Proton invite
in the ICS. For replies, we are going to use the X-PM-PROTON-REPLY field for this.
It has a Boolean value, and it's true when the attendee is replying to a Proton-to-Proton invite.
Other clients can rely on it, but if it's not present they should fallback to checking
via X-PM-SHARED-EVENT-ID as before.

Additionally, although we're still adding the other X-PM fields on replies, we do not
expect to receive them any longer in an ics, since we will later drop them.

CALWEB-2251