Add storybook to the monorepo ci
Change in Components & Styles can change both the final output of
storybook as in what appears in the storybook ui and they are both
direct dependencies and are used in the project so they have the
potential to break build / compilation -> we want to validate for them

Shared, while not being part of what is documented or displayed inside
the storybook ui output, is a direct dependency of the storybook project
and is being directly imported inside it which means that a breaking
change inside of shared could break storybook too. We would like to find
out automatically if it does, so we include shared in the triggers for
storybook.