Fix minicalendar causing jiggling by using aspect-ratio
When having several calendars and a particular screen size (on the threshold
for getting a sidebar), the minicalendar auto-resizing caused an infinite
height adjusting loop and made the sidebar jiggle.

The fix is to use aspect-ratio with @supports where possible to avoid
the infinite computing. The problem would still remain on browser that
do not support aspect-ratio and have this specific amount of content
in the sidebar.

CALWEB-2684