Do not prevent to leave focus from the minicalendar in the sidebar
By default, the minicalendar was preventing the default event action on click.
This is needed when inside a dropdown, but not needed in the sidebar. Actually,
in the sidebar it was preventing to leave focus when clicking on it, which was
generating some unpleasant side effects, basically toolbar buttons remaining focused
while the focus was somewhere else

CALWEB-2623