Migrate to new themes with key maps
The use case here is to reassign user-selected theme from
- old default -> new contrast
- old contrast -> new default
- newly created users -> new default

Wanted to avoid actually migrating these values in the db though
so instead this is moving the themes that need to be redirected to
to a new key (e.g. default goes from 0 -> 6) & then we have a map
to keep track of keys that need to be redirected to newer keys as
they are a redirection (or even a theme that just doesn't exist any
longer in general, like snow/light for example, that's just default now)

For newly created users, agreed with backend devs to set user theme
setting to "6" on creation (again in the endpoint as to not modify the
db schema and avoid a migration)

Another curiosity about this commit might be the introduction of
"PROTON_THEMES_MAP" & "PROTON_THEMES". The reason for this
is that objects which are indexed by numeric keys in JavaScript have
their property appearance order automatically done according to these
numeric keys which is something that went by unnoticed previously but
now causes problems as the keys don't represent the order but rather
are just ids and should be seen as such, therefore a separate array
data-structure is introduced to keep track of the order and just that

CP-2420