Remove inputform css overwrites in storybook index
The input fields in the input stories were rendering incorrectly in
storybook (different than from how they were rendering in the apps).
This seems to be why. Might be a leftover from a transitionary period
for the field migration, not sure, in any case don't believe this is
needed anymore.