Extract optional theme default assignments
Little refactor to see all optional theme variables as well as their
default assignment in one place. There was a need to introduce a
proxy namespace (--optional) in order to be able to do this.
Added benefit, it's not also more evident which variables of the
theme have to be defined and which don't have to be defined.

Co-authored-by: Matthias Margot <margotmatthias@protonmail.com>