Remove appVersionBad condition
Initially copied over from the recently deleted NewVersionTopBanner
it seems that this condition to render the banner when appVersionBad
was true was just there to support the case where the banner was to
be displayed in red.

Since that case is now handled by the BadAppVersionBanner, this
condition should be safely dismissible now.

CP-2069