Fix error in helper normalizing alarm triggers
Due to a mistake in the logic inside the helper, when normalizing an alarm trigger with an integer number of
units expressed in lower units (for instance 60 minutes, 48 hours, 7 days, etc), the normalized trigger would
correspond to a zero duration.
Additionally, when a trigger for a part-day event contains a single component, we do not try to normalize it.

CALWEB-2801