Fix extract finder - remove node_modules
Babel is stupid and it will try to load node_modules (extracted)
when we run ttag.
So to make the find easier to read + remove the scope for bugs
- remove node_modules extracted (thx monorepo, it was legacy)
- remove webpack runtime config extracted

So we can list all the extracted sources and only these files.

Also to ensure it's easier to find, we find _from_ the i18n-js directory