Remove app-root-loader div from verification app
This loader is not really required here as the verification app doesn't
pre-load any data, it's able to render critical path ui immediately
(based on the url parameters).

Having the loader rendered would cause the background of the iframe
to have a color not matching the surrounding UI when embedded, which
would cause a flash of it during load time which looked buggy/laggy.

CP-2489