Generate REPLY from ics envent, not db event
We have been using the db event until now to generate the REPLY ics. In most cases
that is fine since the db event is synchronized with the ics event, but there are
some cases where that won't be the case, like for the Proton-Proton reinvite.
We shouldn't lose anything using the ics one, so let's always use that one.

CALWEB-2745