Handle non-RFC compliant Yahoo REPLY ics
Yahoo does not always respect the RFC when sending a REPLY and sometimes they encode
the partstat in a custom property, X-YAHOO-USER-STATUS, instead of encoding it in the
PARSTAT of the ATTENDEE property. Moreover they do not update DTSTAMP as expected.
We make an effort to recover the partstat from there to display it properly in the widget.

CALWEB-2871