Differentiate between import contact vs add contact in api call
The api is currently unable to tell the difference between the action of
importing contacts (via csv / vcard) vs the action of just adding a contact
because the endpoint for both is the same. This change adds a parameter to
the request through which the client, who is able to tell from the user input,
can communicate to the api whether a contact addition was an import or not.

CP-2526