Inline InvalidVerificationCodeModal in Verification
In order to ensure that the verification application doesn't send
back invalid verification codes to the mobile apps as part of the
completion message "HUMAN_VERIFICATION_SUCCESS" we
need to perform a check for the token's validity from inside the
verification app. This happens via the /users/check endpoint
which is called before actually emitting the success message and
the success message is only emitted if the requets doesn't error
(which it does if the code is invalid).

Problem was, if token submission callback failed (promise reject),
the InvalidVerificationCode Modal would open and modals are a
no-no for mobile embedding as they can't escape their bounding
frame. Therefore, inlined the contents of the modal inside an
additional step in the HumanVerificationForm.

CP-2489