Hide app version name unless sidebar is in vpn settings
Used to be handled with a direct sibling selector in _app-infos.scss
(when app-infos had a direct previous sibling which was the storage span,
it would not render the app-info-name span).

Broke during get-started, when the gift icon was introduced making the gift
icon container the direct preceding sibling of the app-infos. Was replaced with
the condition 'canAddStorage', but that's incorrect.

Instead it just need to be "compact" (meaning that the app name is sr-only) always
unless the sidebar is rendered in vpn.