Clean up and text break
Here, the only problem is if we have a very long email address.
By adding the text-break helper class, the email address will be split
in two or several lines if necessary.
The character '-' will still cause the address to break automatically,
but this is an edge case and the most important thing is to be able
to read the email address entirely.

CALWEB-2604