Improve the resize indicator
We wanted the resize indicator to be on the right edge of the email list,
so we had two solutions:
- place it inside the email list container so that it would be impacted
by the overflow of the email list, so it would follow the scroll bar
if it is present, but the indicator could not overflow this frame...
- or place it outside the overflow container, but it would be shifted
when the scroll bar is present.

We chose the second solution, and to correct this offset, we calculate
the width of the scrollbar and move the indicator with that value.

Co-authored-by: Matthieu Lux <matthieu.lux@proton.ch>

MAILWEB-2816