Set typescript `skipLibCheck` to true
Because project dependencies are bigger, sub dependencies types tend to be
different and conflicts more often as the project grows.

We decided to enable `skipLibCheck` in order to ignore them.

This feature will skip error throwing on every `.d.ts` files encountered in
the project (meaning it's not only related to `node_modules`).

Typescript compilation and check will continue to work as intended. This means
 if source code uses errored `.d.ts` type it'll still throw error.