Use contact signature verification when informing the user if a contact is verified or not
Before this, in the ContactView modal we were always telling the user that the data in the
encrypted fields (like TEL, ADR, ...) was encrypted and verified. That was not too bad since
signature verification errors would be displayed in a very visible banner in the modal, and
contacts were always signed. However, the back-end is creating encrypted (but not signed)
contacts now, so we cannot tell the user that the data in these fields is verified with a
digital signature.

MAILWEB-2897