Don't render payment sections unless user.canPay
Billing, PaymentMethods, Credits, GiftCode & Invoice sections used to be
their own page called "Payment" and that page was rendered on the condition
of user.canPay.

Now, while the permissions on the subsections of the dashboard page should
actually be enough to not physically render these sections in the dom, the
hypothesis is that they are still rendered in react's virtual dom causing
code defined inside to execute and therefore trigger requests for which
there isn't enough permission making the request fail and the page error.

The permission "UPGRADER" is actually the same as user.canPay, which
is why I think the laid out reason causes this issue.

CP-3065