Fix Drive subfolder display
Short story: the Drive subfolder navigation introduced a bug in Mail label display (ellipsis was not working anymore).

We fixed Mail label display, but this created a regression in Drive.
As there was also a potential issue in Drive in "my files" get translated into a loooong text.

So we went for a more radical solution: rework and recheck everything.
In a nutshell:

- rollback Mail fix
- rework Drive root folder display and fix if there is a looong translation
- ellipsis is applied to a lower level

Checked on Account/Drive/Mail/Calendar. TGIF.