Switch to new favicons
Removed usages of svg icons for dynamic favicon replacements via react-helmet
inside of calendar to display the current date. (png was a fallback, however png
is better supported than svgs so we're not using svgs there any more)

We're keeping svgs for the main logo of an app because our build setup uses
this icon to generate additional favicons, which are used in different contexts
e.g. apple-touch-icon.

CP-2418