Calendar model API update. Color and display become member properties
In order to have shared calendars, color and display cannot be calendar-level properties. They need to be member-level properties instead.
This requires changing the management of the list of calendars in the app. Calendars are fetched with members and color and display are
computed as high as possible in the component tree. Updates to calendar or member data can come from different places via the event loop.
The use of model event listeners is deprecated for calendar, and instead a custom listener hook should be used.

CALWEB-3295