Close auth modals on error
Leave them open for authentication errors to allow the user to try
again. If the authentication succeeds but the request errors out,
close it as it probably means the original API request failed for
some reson.

Remove the inline authentication errors as they're not handled properly
by clearing them on input.