Introduce the concept of centralised theme colors
Generally speaking (unless we want to restrict it for particular
use-cases like in ButtonGroup), components should support the
same group of colors throughout the component system or at least
pick from the same palette.

So far this palette was only defined in CSS and on the JS/TS side
hard-coded per-component. This change introduces a
single-source-of-truth for the generally available theme colors.