Delay theme data query
The challenge component is a child to unauthenticated. Both of them rely
on useLayoutEffect. The challenge component's useLayoutEffect triggers
before unauthenticated, causing it to fetch an old theme. This delays
the fetching until after the iframe has loaded which in the real world
is a good enough assumption to always happen.