Fix potentially bad payload for calendar reset
There are two ways of getting the list of calendar to be reset. One is from the normal list of calendars,
using the flags. The other one is through the specific /keys/reset route. Both lists should contain the same
calendars, but not necessarily in the same order. In the cases where the order is different, they payload
we would generate wouldn't be correct and calendars wouldn't be reset.

CALWEB-3536