Add a side-panel to Mail, used to display the Calendar app MAILWEB-3124
In the first version, the side-panel allows to consult Calendar, create, edit, and delete events, but also to open events in the side-panel from an email.
Calendar app is loaded inside an iframe. The communication between the parent and child app is done with a postMessage(), and we specifically limit the communication between Proton Mail and Proton Calendar, so no one can listen exchanged data.

Co-Authored-By: @econdepe
Co-Authored-By: @sergonius
Co-Authored-By: @mlux
Co-Authored-By: @nhoffmann