Remove faker.js dependencies
faker.js is not maintained anymore. Its developer actually updated corrupted the library on purpose
[https://www.theverge.com/2022/1/9/22874949/developer-corrupts-open-source-libraries-projects-affected]
so it's not safe to keep the dependency in the code in case someone updates it by mistake. We barely used it anyway

The test-data-bot is also updated to a version which does not rely on faker.js