Addm minimumCycle search param to SubscriptionData
`minimumCycle` defines the minimum length cycle option to display.
If this is greater than the max cycle then only the max cycle is
shown. If it is not defined then all cycles are shown.

When signing up from the proton.me pricing page, any cycles that
are less time than the minimumCycle passed in the query params is not shown.

CP-4367