Encrypt in binary to avoid OpenPGP normalization with trailing spaces
If we encrypt text (instead of Uint8Array data), we have to tell OpenPGP how to normalize trailing spaces.
As it was the usual practice before to strip them, we have continued doing it (when signing, it is instead
important to be backwards-compatible).

This poses problems in Calendar because of the folding convention in an ICS, where trailing spaces have a meaning.
For certain long event titles, locations or descriptions, if a whitespace hits the end of a folded line in
the ICS, OpenPGP would eat it and it would disappear from the event after saving it.

The best option for us is to transform the ICS into a UInt8Array before encrypting it, so that OpenPGP
preserves the exact format and does not apply any normalization (like changing the format of end of lines from
\r\n to \n, that is only relevant in the Mail context).

Signing of clear-text cards is left untouched for backwards compatibility.

CALWEB-4019