Make Proton logo no clickable when hasDecoration prop is false
If hasDecoration prop is false, then we do not want to distract the
user. In this case the logo shoul not be clickable, taking them away
from the signup flow.

This change affects anywhere the hasDecoration prop is false, including
the oauth confirmation, extension login confirmation, 2fa or mailbox
during sign in.

CP-5082