Catch errors getting public keys and log more errors reading calendar events
We catch errors when getting public keys (needed to verify signatures of invites)
since it's better to not verify than to display an error in the popover.

Additionally, we add new logs to debug some current network instability issues we are seeing.
Decryption errors are not logged. No sensitive data is logged either

CALWEB-4115