Adapt app to change of alarms in the API architecture
Personal events will be deprecated in favor of a new field in the CalendarEvent object: Notifications. We need to live for a while with both.

* CalendarEvent contains a property IsPersonalMigrated. When that is true, we should extract alarms from the Notifications field.
Otherwise extract from PersonalEvents as we were doing up to now.

* In the new architecture, Notifications = null means that the event has default calendar notifications (empty notifications would be signaled by
Notifications = []).

* The signature on the personal events are lost.

The reason for this change is that the previous architecture was devised with shared addresses in mind, but that is not happening anytime soon.
The new architecture is faster (no need to write/retrieve blobs), and it will allow us to align alarms UX for personal, shared and subscribed calendars.

CALWEB-4004