Remove unused code
The "organizer disabled alert" would never show up. If the organizer can't send email, we don't allow her to edit
the event at all. So it's not possible to ever hit the EventForm in such cases. It was possible in the past, that's
the reason why the code was there (it should have been removed with CALWEB-2267)