Sign and verify all calendar cards as binary data
OpenPGP offers two modes of signing text:
 * As binary data: after converting the text into binary data, the data is signed as is.
 * As text data: the text is first "normalized" according to an OpenPGP spec, then signed.
   The spec always normalizes end of lines "\n" to "\r\n". Additionally, we used to strip
   trailing spaces (according to OpenPGP spec, this is to be done only for "cleartext messages"
   (special type of OpenPGP signed message), but we used to apply the same normalisation whenever
   we signed text).

Not so long ago we started to encrypt and sign encrypted cards as binary data. Non-encrypted cards were however always signed as text data.
In the Calendar context it makes little sense to apply the OpenPGP normalization which is meant only for a Mail context. We therefore would like
to start using only binary signatures.