Bring back batching for event reading
It turns out that there's a bug in Chromium (see https://github.com/GoogleChrome/workbox/issues/2528 for
multiple references to it) that causes an exhaustion of the browser resources when launching many simultaneous
HTTP2 requests (not an issue for HTTP1 since in that case simultaneous requestst are capped at 6), making
requests over the quota fail with a net:ERR_INSUFFICIENT_RESOURCES error. Only Chromium-based browser have this bug.
The number of max simultaneous requests allowed seems to be variable. There's probably not a hard cap on it,
but rather a cap on the memory that the browser can use. In our tests we saw the error at around ~1000 requests.

CALWEB-4446