Preserve file modification timestamps in ZIP archives
Currently, files in downloaded ZIP archives have current time as a
modification time, even though the modification time is preserved on the
server side when uploading files. That seems like a regression to me,
since a sequence of upload-download effectively erases this metadata,
which is unexpected, inconvenient at best and may result in a complete loss.

I can imagine (1) moving files from some place to ProtonDrive, then (2) moving
these files or some subset of them out of ProtonDrive. No backups from (1) and
erase ProtonDrive's copy at (2) without checking timestamps during the move
results in a loss. If one happened to have backups and ProtonDrive's copy is
earsed, that may be a partial loss if files have been modified in ProtonDrive
between (1) and (2). If ProtonDrive's copy is accessible during or after (2)
and there are no backups from (1) or there were changes in ProtonDrive between
(1) and (2), prepare for manual timestamp mangling which may as well be
infeasible depending on the number of files.

The timestamp is currently set by `client-zip`, where `new Date()` is a default
when input is `ReadableStream<Uint8Array>` and no `lastModified` provided.
Since `fileModifyTime` is already present on a `DecryptedLink`, just pass that
through to `archiveGenerator` and set `lastModified`.