Allow admins with 2fa to change members password
Admins with 2fa could not change a members password through the user
table quick actions as the unlock password scope call was using an
expired 2fa token (the token was used by the `authMember` call).

To fix this without creating a new endpoint, the change password call
now prompts a second auth modal. This is not ideal and will be changed
in the future to only prompt once.