Avoid unnecessary calls to get contact emails in Calendar
Because we were not passing the contactEmailsMap to the hook getting the verification keys,
needed to verify events by useGetCalendarEventRaw, the verification hook was making unnecessary
calls to get contact emails. Those are not needed as the ContactEmails model is preloaded by all apps.

The ideal solution would be to get the map from a hook as well, but each app is computing this
map in a different way. Mail computes it for the Redux store, while Calendar and Account use
a dedicated hook useContactEmailsCache. Drive is not ever computing this map, but they are not
using the useVerificationPreferences for the moment.

Therefore, in these hunger games with no Core team coordinating apps, we're forced to go for a dirty
solution where the map is passed directly to useGetCalendarEventRaw when that is instantiated.

CALWEB-4600