Avoid double re-render that would block the event reader in the UI
A nasty race condition was getting triggered by calling the model event manager
right before re-rendering the UI upon event change actions.
The race condition comes about because the upserts from the model event and the event
change actions happen near simultaneously, and one blocks the rendering of the other.

This MR is not a proper solution, just a temporary patch