Resolve race condition in useCalendarsEventsReader
There was a race condition in the code where the promises to read (decrypt) an event could be launched, and while they were resolving an event loop event
would come that would trigger a re-run of the useMemo of useCalendarEvents, which in turn would re-run the useEffect in useCalendarsEventsReader.
By the time the original promise would resolve, isActive = false (because of the useEffect re-run), which would prevent the code from reaching
the re-render with the event result. This results in the event blinking forever (or more like until the useCalendarsEventsReader useEffect would.
be triggered again).

CALWEB-4648