Fix reply-to/forward of E2EE forwarded message with attachments
When creating the draft, the attachment key packets need to be decrypted
with the forwardee key, which requires setting a special flag to be used.
This flag was not set on draft creation.
This commit adds `MessageDraftFlags.originalMessageFlags` to make the
forwarding information accessible to `createNewDraft`.
After the draft has been created, and the attachment keys re-encrypted, the
information can be discarded; hence, it does not need BE-side storage.