Handle duplicate ids for calendar events
Calendar events can share the same event ID as long as they live in different calendars.
Such duplicates will occur in production once the calendar DB gets split in shards.

Previously we handled this possibility in some places, but not all. Most notably, the UI
would get confused by such duplicates as we were relying on the event id for several identifications
(like which event to show in the popover).

I made the id of the CalendarViewEvent object unique (by combining calendar id and event id into a unique id)
so that the code does not need to change much.

CALWEB-4702