Handle properly deletion of single cancellation attendee-side when all events are cancelled
When deleting a single deletion in a cancelled series, there's no need to update the parent event
with an EXDATE, as the EXDATE was already present. The API will complain if we try to do so.

Notice in this case there's a competing RECURRENCE-ID and an EXDATE. This was not possible before,
so the code needs to be adapted.

CALWEB-4755