Handle session recovery insecure state when less than an hour remaining
Update `useSessionRecoveryInsecureTimeRemaining` to handle times less
than an hour. It automatically updates the diff at specific intervals
depending on the time remaining.

When there is an hour or less remaining then the diff is updated every
minute, when there is a minute or less remaining then the diff is
updated every second.

CP-6996