Harmonize list display on inputs
Chrome is adding a quite ugly caret, and badly positionned.

We considered replacing it by a suffix like a caret icon for the select.
However, after some tries, we discovered that it can be misleading:

- looks like a select but it's not
- caret is displayed even if list is empty (which is more misleading, user won't understand that nothing happens in this case)

So we went to hide it only, because we have no way to know the state of this list, as it's browser-based (and not state/pseudo-element to know if the list is empty or if the list is opened/closed)

UXE-787