Fix timezone field display
First fix of `h-full` was for mobile (to align fields)`, and it was okay on desktop. However, we missed the timezone case.

Problem: `h-full md:h-auto` looked like a simple fix, but is not working, because we must NOT set up `height` at all at higher breakpoints (browser does not care about, and no other value works).

- Applied `h-full` class only when on mobile with useActiveBreakpoint.

UXE-789