Remove Safari autocomplete on Composer
Safari was triggering an autocomplete even with autocomplete="off"… because.

Looking at https://stackoverflow.com/questions/22661977/disabling-safari-autofill-on-usernames-and-passwords not all workarounds are working.

The CSS one is not working (I did try a lot to make it work). The fake input neither. `type=search` is working but breaking layout.

However, adding `name="search"` to the input seems to do the job, plus disabling autocorrect and spellcheck.

Bonus: removed an aria-hidden that was not useful.

MAILWEB-4797