Mail tests: fix fake-timer-related timeouts and other race conditions when using the v6 CryptoAPI
The issues are linked to the fact that v6 uses SubtleCrypto for all supported operations,
unlike v5 which bypassed it for small amounts of data.
Fake timers do not work with SubtleCrypto, which always runs in 'real time'.
A few other race conditions popped up, likely linked to timing or async resolution
differences with the new version.
The affected tests have been changed to wait for specific side effects to complete,
when possible, rather than relying on arbitrary timeouts.