Do not growl for escape block errors
In a few places we were rejecting a promise with the intention of escaping a block. The intent was achieved
by a coincidence, as the catch code was trying to extract an error message, and logging an error to Sentry
when not being able to do so. Some recent code fixed the Sentry part, but introduced growling an "unknown error".
The solution proposed is to mark the intent of the "escape block" error as an error class

CALWEB-4812