Improve day button vocalization
Day buttons were vocalized "Mon 29" and then the title ("Monday, January 29th, 2024") when focusing it. So we have a double vocalization, with a first part not always explicit.

Moreother, for users that are printing the list of buttons (in NVDA for example), it's only the first part which will be printed out.

- Added full date in `aria-label`, so both cases are solved (no double vocalization and giving full date for both)

UXE-683