Optional override for chargebeeEnabled in payment facade
It helps to handle the case when user tries to upgrade to Pass B2B via Pass Signup page while it's already a forced Chargebee user. Before the fix, the user was handled as allowed Chargebee user, and as a result they had an in-house credit card form for Pass B2B upgrade.