TimeOutOfSyncTopBanner: do not warn on stale server time update
The server time update might be stale if e.g. the device has been asleep/idle,
and it's processing the update after a delay.
The event loop runs every 30s, so we expect the server time to be updated at
least with that frequency (with a margin of a few seconds)