Add support for v6 active address keys (reading only)
We need to store address active keys split by version, to make it easier
to enforce the required orderings and normalizations on
primary key changes.
A v4 primary address key is always required, regardless of
the presence of v6 keys. This is for backwards compatibility,
since address keys are fetched by other users, and not all clients
are currently compatible with v6 keys.
A v6 primary address key is not required, even if v6 address keys
are present. This is because e2ee forwarding cannot work with v6
keys, hence users are allowed to mark their v6 primary address key
as non-primary to be able to set up forwarding.

Instead, for account keys (also known as user keys),
there is no need to split the internal logic to keep track of the key version.
In fact, there is always a single primary key (v4 or v6), since backwards compatibility
is not a concern, as these keys are for personal use only.