Optimize event manager bootstrap
This allows us to avoid waiting for the latest event id in bootstrap,
and we'll fetch it dynamically instead. This has the drawback of it
potentially being wrong if the initial call fails, and the user
performs actions since it'll retry getting the latest event id after
the action. However we want to optimize for the standard scenario
over rare occurrence.