CryptoAPI: fix `isE2EEForwardingKey` to support keys without valid encryption subkeys
The function used to mistakenly check only for the absence of valid encryption keys.
This would work in practice as forwarding subkeys are not valid for encryption, and
standard private address keys are expected to have valid encryption subkeys.
However, it caused issues in a specific account setup where the user imported
an expiring key, which was then wrongly considered to be a forwarding key, which prevented
the user from generating a valid SKL as a result (the FE filtered out the key on SKL generation,
while the BE expected the key to be included).

CRYPTO-164 (follow up)