`resignSKLWithPrimaryKey`: skip if primary keys are newer than the old SKLs' signature timestamp
If the keys are freshly generated or imported, they may not be
valid for signing at the required timestamp, and signing would fail.

Since new keys do not appear in the old SKLs, re-signing is not fundamental,
as the chain of events can be reconstructed looking at the SKL changes
rather than the SKL timestamps.