`verifySKLSignature`: allow any of the provided `verificationKeys` to verify the SKL
Verification cannot be restricted to using the primary keys declared in the SKL,
because for older SKLs those keys may no longer be available among the address verification
keys (e.g. because the keys were deleted or marked as compromised by the user).

Also, even if the SKL includes a PQC signature, we don't mandate that it verifies over
the non-PQC one; in fact, at the moment we do not provide PQC KT protections against the BE:
if non-PQC signature forgery were possible, nothing currently prevents the BE from generating
a new SKL where the PQC key isn't primary anymore.
To tackle this issue, we will have to enforce PQC verification for all KT users and SKLs.