Mail BYOE: hide lock icons when sending to a mixed set of recipients that include non-e2ee ones
As the message will ultimately be sent externally to all recipients, with encryption disabled.

If all external recipients have e2ee enabled (without EO), lock icons remain unchanged, but the BE will error as the message
cannot be sent.
If EO is enabled, encryption remains in place as the message is still delivered internally only.

In all cases the send preferences remain unchanged, and won't match the lock icons, as the BE still enforces sending
encrypted packages for recipients (both internal and external) where e2ee is expected.